split out tests into better structure
- old was single file
- new is nice and modular
