+++ tests/api/subscriptions.rs
@@ -1,0 +1,63 @@
+use crate::helpers::spawn_app;
+
+#[actix_rt::test]
+async fn subscribe_returns_a_200_for_valid_form_data() {
+    let app = spawn_app().await;
+    let body = "name=col%20man&email=colo%40col.com";
+
+    let response = app.post_subscriptions(body.into()).await;
+
+    assert_eq!(200, response.status().as_u16());
+
+    let saved = sqlx::query!("SELECT email, name FROM subscriptions",)
+        .fetch_one(&app.db_pool)
+        .await
+        .expect("Failed to fetch saved subscription.");
+
+    assert_eq!(saved.email, "colo@col.com");
+    assert_eq!(saved.name, "col man");
+}
+
+#[actix_rt::test]
+async fn subscribe_returns_a_400_for_missing_data() {
+    let app = spawn_app().await;
+
+    let test_cases = vec![
+        ("name=col%20man", "missing the email"),
+        ("email=colo%40col.com", "missing the name"),
+        ("", "missing both name and email"),
+    ];
+
+    for (invalid_body, error_message) in test_cases {
+        let response = app.post_subscriptions(invalid_body.into()).await;
+
+        assert_eq!(
+            400,
+            response.status().as_u16(),
+            "The API did not fail with 400 Bad Request when payload was {}",
+            error_message
+        );
+    }
+}
+
+#[actix_rt::test]
+async fn subscribe_returns_a_400_when_fields_are_present_but_invalid() {
+    let app = spawn_app().await;
+
+    let test_cases = vec![
+        ("name=&email=col@check", "empty name"),
+        ("name=Colman&email=", "empty email"),
+        ("name=colo&email=im-not-an-email", "invalid email"),
+    ];
+
+    for (body, description) in test_cases {
+        let response = app.post_subscriptions(body.into()).await;
+
+        assert_eq!(
+            400,
+            response.status().as_u16(),
+            "The API did not return a 400 Bad Request when the payload was {}.",
+            description
+        );
+    }
+}
